Updated FlickrStore to latest Flickr API (via FlickrJ)

+++ trunk/FlickrStore/src/be/vanvlerken/bert/flickrstore/login/FlickrUser.java
@@ -14,9 +14,9 @@ import com.aetrion.flickr.contacts.OnlineStatus;
 public interface FlickrUser
 {
     public String getAwayMessage();
-    public int getBandwidthMax();
-    public int getBandwidthUsed();
-    public int getFilesizeMax();
+    public long getBandwidthMax();
+    public long getBandwidthUsed();
+    public long getFilesizeMax();
     public int getIconServer();
     public String getId();
     public String getLocation();
